+++ lib/src/network/protocol/mod.rs
@@ -1,8 +1,8 @@
 pub mod http;
 pub mod pipe;
-pub mod tls;
+pub mod openssl;
 
-pub use self::tls::TlsHandshake;
+pub use self::openssl::TlsHandshake;
 pub use self::pipe::Pipe;
 pub use self::http::{Http,StickySession};
